Refactored credit_card_payments.py to installments.py & adjusted whatsapp message if there are no installments
