+++ main.py
@@ -7,7 +7,7 @@ from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 
 import automation
-import credit_card_payments as ccp
+import installments as ccp
 import data_services
 import whatsapp
 from DTO import WhatsappMessageDTO
